Use Actions for DamageScript
