+++ projects/Layer 0 - Domain/Eevee/vampirekiller.eevee/statements/schemas/DamageSchema.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
+using vampirekiller.eevee.statements;
 
 namespace VampireKiller.eevee.vampirekiller.eevee.statements.schemas;
